Lógica do Sorteio (#8)
* Ajustes nos scripts

* Aplicando lógica do sorteio

* Ajustando nomenclatura fks

+++ src/WebAPI/AlienTorpedoAPI/Models/GrupoUsuario.cs
@@ -6,7 +6,7 @@ namespace AlienTorpedoAPI.Models
     public partial class GrupoUsuario
     {
         public int CdUsuario { get; set; }
-        public short CdGrupo { get; set; }
+        public int CdGrupo { get; set; }
         public short? NrVoto { get; set; }
 
         public virtual Grupo CdGrupoNavigation { get; set; }
